+++ src/stores/Scheme.js
@@ -5,10 +5,14 @@ export const useSchemeStore = defineStore('scheme', {
     scheme: {
       // chromatic
       hue: -15,                               // int [0-360]
+      hueSet: 'standard',                     // enum ('uno', 'duo', 'trio', 'standard')
+      hueDistance: 0,                         // int [0-45]
       degrees: [0, 60, 120, 180, 240, 300],   // [int, int, int, int, int, int]
       saturation: 50,                         // float [0-100]
+      saturationRange: 0,                     // int [0-50]
       normalChromaticLightness: 50,           // float [0-100]
       brightChromaticLightness: 75,           // float [0-100]
+      lightnessRange: 0,                      // int [0-30]
       // achromatic
       normalBlackLightness: 0,                // float [0-100]
       brightBlackLightness: 12.5,             // float [0-100]